Adds requirement 76 to the test file of requirement 75
